Refactoring Domain Layer
