+++ app/src/main/java/com/unsa/suppliers/domain/auth/LoginUseCase.kt
@@ -1,14 +1,14 @@
 package com.unsa.suppliers.domain.auth
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.auth.LoginRequest
 import com.unsa.suppliers.data.dtos.auth.LoginResponse
+import com.unsa.suppliers.data.repositories.AuthRepository
 import javax.inject.Inject
 
 class LoginUseCase @Inject constructor (
-    private val repository: SupplierRepository
+    private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(loginRequest: LoginRequest): LoginResponse? {
-        return repository.attemptLogin(loginRequest)
+        return authRepository.attemptLogin(loginRequest)
     }
 }
