+++ app/src/main/java/com/unsa/suppliers/domain/auth/RegisterUseCase.kt
@@ -1,14 +1,14 @@
 package com.unsa.suppliers.domain.auth
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.auth.UserRequest
 import com.unsa.suppliers.data.dtos.auth.UserResponse
+import com.unsa.suppliers.data.repositories.AuthRepository
 import javax.inject.Inject
 
 class RegisterUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val authRepository: AuthRepository
 ) {
     suspend operator fun invoke(userRequest: UserRequest): UserResponse? {
-        return supplierRepository.attemptRegister(userRequest)
+        return authRepository.attemptRegister(userRequest)
     }
 }
