+++ app/src/main/java/com/unsa/suppliers/domain/main/categories/DeleteCategoryUseCase.kt
@@ -1,12 +1,12 @@
 package com.unsa.suppliers.domain.main.categories
 
-import com.unsa.suppliers.data.SupplierRepository
+import com.unsa.suppliers.data.repositories.CategoryRepository
 import javax.inject.Inject
 
 class DeleteCategoryUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val categoryRepository: CategoryRepository
 ) {
     suspend operator fun invoke(id: Int) {
-        supplierRepository.attemptDeleteCategory(id)
+        categoryRepository.attemptDeleteCategory(id)
     }
 }
