+++ app/src/main/java/com/unsa/suppliers/domain/main/categories/EnableCategoryUseCase.kt
@@ -1,12 +1,12 @@
 package com.unsa.suppliers.domain.main.categories
 
-import com.unsa.suppliers.data.SupplierRepository
+import com.unsa.suppliers.data.repositories.CategoryRepository
 import javax.inject.Inject
 
 class EnableCategoryUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val categoryRepository: CategoryRepository
 ) {
     suspend operator fun invoke(id: Int) {
-        supplierRepository.attemptEnableCategory(id)
+        categoryRepository.attemptEnableCategory(id)
     }
 }
