+++ app/src/main/java/com/unsa/suppliers/domain/main/categories/GetAllCategoriesUseCase.kt
@@ -1,13 +1,13 @@
 package com.unsa.suppliers.domain.main.categories
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.categories.CategoryResponse
+import com.unsa.suppliers.data.repositories.CategoryRepository
 import javax.inject.Inject
 
 class GetAllCategoriesUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val categoryRepository: CategoryRepository
 ) {
     suspend operator fun invoke(): MutableList<CategoryResponse> {
-        return supplierRepository.attemptGetCategories() ?: mutableListOf()
+        return categoryRepository.attemptGetCategories() ?: mutableListOf()
     }
 }
