+++ app/src/main/java/com/unsa/suppliers/domain/main/categories/GetCategoryByIdUseCase.kt
@@ -1,13 +1,13 @@
 package com.unsa.suppliers.domain.main.categories
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.categories.CategoryResponse
+import com.unsa.suppliers.data.repositories.CategoryRepository
 import javax.inject.Inject
 
 class GetCategoryByIdUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val categoryRepository: CategoryRepository
 ) {
     suspend operator fun invoke(id: Int): CategoryResponse {
-        return supplierRepository.attemptGetCategoryById(id) ?: CategoryResponse(0, "Error", "Error")
+        return categoryRepository.attemptGetCategoryById(id) ?: CategoryResponse(0, "Error", "Error")
     }
 }
