+++ app/src/main/java/com/unsa/suppliers/domain/main/countries/CreateCountryUseCase.kt
@@ -1,13 +1,13 @@
 package com.unsa.suppliers.domain.main.countries
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.countries.CountryRequest
+import com.unsa.suppliers.data.repositories.CountryRepository
 import javax.inject.Inject
 
 class CreateCountryUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val countryRepository: CountryRepository
 ) {
     suspend operator fun invoke(countryRequest: CountryRequest) {
-        supplierRepository.attemptCreateCountry(countryRequest)
+        countryRepository.attemptCreateCountry(countryRequest)
     }
 }
