+++ app/src/main/java/com/unsa/suppliers/domain/main/countries/EnableCountryUseCase.kt
@@ -1,12 +1,12 @@
 package com.unsa.suppliers.domain.main.countries
 
-import com.unsa.suppliers.data.SupplierRepository
+import com.unsa.suppliers.data.repositories.CountryRepository
 import javax.inject.Inject
 
 class EnableCountryUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val countryRepository: CountryRepository
 ) {
     suspend operator fun invoke(id: Int) {
-        supplierRepository.attemptEnableCountry(id)
+        countryRepository.attemptEnableCountry(id)
     }
 }
