+++ app/src/main/java/com/unsa/suppliers/domain/main/countries/GetAllCountriesUseCase.kt
@@ -1,13 +1,13 @@
 package com.unsa.suppliers.domain.main.countries
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.countries.CountryResponse
+import com.unsa.suppliers.data.repositories.CountryRepository
 import javax.inject.Inject
 
 class GetAllCountriesUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val countryRepository: CountryRepository
 ) {
     suspend operator fun invoke(): MutableList<CountryResponse> {
-        return supplierRepository.attemptGetCountries() ?: mutableListOf()
+        return countryRepository.attemptGetCountries() ?: mutableListOf()
     }
 }
