+++ app/src/main/java/com/unsa/suppliers/domain/main/countries/GetCountryByIdUseCase.kt
@@ -1,13 +1,13 @@
 package com.unsa.suppliers.domain.main.countries
 
-import com.unsa.suppliers.data.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.countries.CountryResponse
+import com.unsa.suppliers.data.repositories.CountryRepository
 import javax.inject.Inject
 
 class GetCountryByIdUseCase @Inject constructor (
-    private val supplierRepository: SupplierRepository
+    private val countryRepository: CountryRepository
 ) {
     suspend operator fun invoke(id: Int): CountryResponse {
-        return supplierRepository.attemptGetCountryById(id) ?: CountryResponse(0, "Error", "Error")
+        return countryRepository.attemptGetCountryById(id) ?: CountryResponse(0, "Error", "Error")
     }
 }
