+++ app/src/main/java/com/unsa/suppliers/domain/main/suppliers/DeleteSupplierUseCase.kt
@@ -1,6 +1,6 @@
 package com.unsa.suppliers.domain.main.suppliers
 
-import com.unsa.suppliers.data.SupplierRepository
+import com.unsa.suppliers.data.repositories.SupplierRepository
 import javax.inject.Inject
 
 class DeleteSupplierUseCase @Inject constructor (
