+++ app/src/main/java/com/unsa/suppliers/domain/main/suppliers/GetSupplierByIdUseCase.kt
@@ -1,6 +1,6 @@
 package com.unsa.suppliers.domain.main.suppliers
 
-import com.unsa.suppliers.data.SupplierRepository
+import com.unsa.suppliers.data.repositories.SupplierRepository
 import com.unsa.suppliers.data.dtos.main.suppliers.SupplierResponse
 import javax.inject.Inject
